+++ software/tools/fau-calibration.c
@@ -66,49 +66,40 @@ static int fau_calibration_read(char *path, struct fa_calib *calib,
 	return ret;
 }
 
-static void fau_calibration_dump_stanza(struct fa_calib_stanza *stanza)
+static void fau_calibration_dump_stanza(const struct fa_calib_stanza *stanza)
 {
-	fprintf(stdout, "  temperature: %f C\n",
-		stanza->temperature * 0.01);
-	fprintf(stdout, "  gain: [0x%04"PRIx16", 0x%04"PRIx16", 0x%04"PRIx16", 0x%04"PRIx16"]\n",
-		stanza->gain[0],
-		stanza->gain[1],
-		stanza->gain[2],
-		stanza->gain[3]);
-	fprintf(stdout, "  offset: [0x%04"PRIx16", 0x%04"PRIx16", 0x%04"PRIx16", 0x%04"PRIx16"]\n",
-		stanza->offset[0],
-		stanza->offset[1],
-		stanza->offset[2],
-		stanza->offset[3]);
+	int i;
+	fprintf(stdout, "    temperature: %f C\n",
+			le16toh(stanza->temperature) * 0.01);
+	fputs("    calibration:\n", stdout);
+	for (i = 0; i < 4; ++i)
+		fprintf(stdout, "      - {channel: %d, gain: 0x%04"PRIx16", offset: 0x%04"PRIx16"}\n",
+				i + 1, (uint16_t)le16toh(stanza->gain[i]),
+				(uint16_t)le16toh(stanza->offset[i]));
 }
 
 /**
  * Print calibration data on stdout in humand readable format
  * @calib: calibration data
  */
-static void fau_calibration_dump_human(struct fa_calib *calib)
+static void fau_calibration_dump_human(const struct fa_calib *calib)
 {
-	uint16_t *data16 = (uint16_t *)calib;
-	int i;
-
-	/* Fix endianess */
-	for (i = 0; i < sizeof(*calib) / sizeof(uint16_t); ++i)
-		data16[i] = le16toh(data16[i]);
-
-	fputs("ADC Range 10V\n", stdout);
+	fputs("ADC:\n", stdout);
+	fputs("  - Range: 100mV\n", stdout);
+	fau_calibration_dump_stanza(&calib->adc[FA100M14B4C_RANGE_100mV]);
+	fputs("  - Range: 1V\n", stdout);
+	fau_calibration_dump_stanza(&calib->adc[FA100M14B4C_RANGE_1V]);
+	fputs("  - Range: 10V\n", stdout);
 	fau_calibration_dump_stanza(&calib->adc[FA100M14B4C_RANGE_10V]);
-	fputs("DAC Range 10V\n", stdout);
+
+	fputs("DAC:\n", stdout);
+	fputs("  - Range: 100mV\n", stdout);
+	fau_calibration_dump_stanza(&calib->dac[FA100M14B4C_RANGE_100mV]);
+	fputs("  - Range: 1V\n", stdout);
+	fau_calibration_dump_stanza(&calib->dac[FA100M14B4C_RANGE_1V]);
+	fputs("  - Range: 10V\n", stdout);
 	fau_calibration_dump_stanza(&calib->dac[FA100M14B4C_RANGE_10V]);
 
-	fputs("ADC Range 1V\n", stdout);
-	fau_calibration_dump_stanza(&calib->adc[FA100M14B4C_RANGE_1V]);
-	fputs("DAC Range 1V\n", stdout);
-	fau_calibration_dump_stanza(&calib->dac[FA100M14B4C_RANGE_1V]);
-
-	fputs("ADC Range 100mV\n", stdout);
-	fau_calibration_dump_stanza(&calib->adc[FA100M14B4C_RANGE_100mV]);
-	fputs("DAC Range 100mV\n", stdout);
-	fau_calibration_dump_stanza(&calib->dac[FA100M14B4C_RANGE_100mV]);
 	fputc('\n', stdout);
 }